feat: Remove ember-animated and refactor <Menu> animation
Ember-css-transitions seems easier to use and is a bit more... clearly
contemporary compared to ember-animated / liquid-fire. However, this
also involves reimagining how the menu appears. Instead of pushing the
main content out of the way, it now slides over it from offscreen. I
think this model will help when transitioning to non-mobile design.